+++ src/bosh-alicloud-cpi/action/call_context.go
@@ -6,10 +6,7 @@ package action
 import (
 	"bosh-alicloud-cpi/alicloud"
 	"bosh-alicloud-cpi/registry"
-	"fmt"
-	"strings"
 
-	bosherr "github.com/cloudfoundry/bosh-utils/errors"
 	boshlog "github.com/cloudfoundry/bosh-utils/logger"
 )
 
@@ -38,22 +35,23 @@ func NewCallContext(input []byte, logger boshlog.Logger, config alicloud.Config)
 	}
 }
 
-func cleanString(s string) string {
-	s = strings.Replace(s, "\n", "", -1)
-	s = strings.Replace(s, "\t", "", -1)
-	return s
-}
-
-func (c CallContext) Errorf(msg string, args ...interface{}) error {
-	s := "input=`" + cleanString(c.Input) + "` " + fmt.Sprintf(msg, args...)
-	return bosherr.Error(s)
-}
-
-func (c CallContext) WrapError(err error, msg string) error {
-	return bosherr.WrapErrorf(err, "input=`%s` message=%s", cleanString(c.Input), msg)
-}
-
-func (c CallContext) WrapErrorf(err error, msg string, args ...interface{}) error {
-	s := "input=`" + cleanString(c.Input) + "` " + fmt.Sprintf(msg, args...)
-	return bosherr.WrapError(err, s)
-}
+//Todo The next method has been deprecated from 29.0.0 and use bosherr methods instead. The next version should be removed.
+//func cleanString(s string) string {
+//	s = strings.Replace(s, "\n", "", -1)
+//	s = strings.Replace(s, "\t", "", -1)
+//	return s
+//}
+//
+//func (c CallContext) Errorf(msg string, args ...interface{}) error {
+//	s := "input=`" + cleanString(c.Input) + "` " + fmt.Sprintf(msg, args...)
+//	return bosherr.Error(s)
+//}
+//
+//func (c CallContext) WrapError(err error, msg string) error {
+//	return bosherr.WrapErrorf(err, "input=`%s` message=%s", cleanString(c.Input), msg)
+//}
+//
+//func (c CallContext) WrapErrorf(err error, msg string, args ...interface{}) error {
+//	s := "input=`" + cleanString(c.Input) + "` " + fmt.Sprintf(msg, args...)
+//	return bosherr.WrapError(err, s)
+//}